agents-portal/agents/docstoteles/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
python-multipart==0.0.6
openai==1.58.1
python-dotenv==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
chromedriver-autoinstaller==0.6.4
langchain==0.1.0
langchain-openai==0.0.5
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
